requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_ipyw_jlab_nb_ext_conf=0.1.0=py38_0
_libgcc_mutex=0.1=main
absl-py=0.12.0=py38h06a4308_0
aiohttp=3.7.4=py38h27cfd23_1
anaconda=2020.11=py38_0
anaconda-client=1.7.2=py38_0
anaconda-navigator=1.10.0=py38_0
anaconda-project=0.8.4=py_0
backcall=0.2.0=py_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=pyhd3eb1b0_0
backports.shutil_get_terminal_size=1.0.0=py38_2
backports.tempfile=1.0=pyhd3eb1b0_1
backports.weakref=1.0.post1=py_1
conda=4.9.2=py38h06a4308_0
conda-build=3.20.5=py38_1
conda-env=2.6.0=1
conda-package-handling=1.7.2=py38h03888b9_0
conda-verify=3.4.2=py_1
contextlib2=0.6.0.post1=py_0
future=0.18.2=py38_1
get_terminal_size=1.0.0=haa9412d_0
idna=2.10=py_0
imageio=2.9.0=py_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
iniconfig=1.1.1=py_0
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38hb070fc8_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
isort=5.6.4=py_0
jpeg=9b=h024ee3a_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab=2.2.6=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
keras=2.4.3=0
keras-base=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
libarchive=3.4.2=h62408e4_0
libcurl=7.71.1=h20c2e04_1
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
liblief=0.10.1=he6710b0_0
libllvm10=10.0.1=hbcb73fb_5
markdown=3.3.4=py38h06a4308_0
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mock=4.0.2=py_0
more-itertools=8.6.0=pyhd3eb1b0_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.1.0=py38_0
msgpack-python=1.0.0=py38hfd86e86_1
multidict=5.1.0=py38h27cfd23_2
multipledispatch=0.6.0=py38_0
navigator-updater=0.2.1=py38_0
networkx=2.5=py_0
ninja=1.10.2=py38hff7bd54_0
nltk=3.5=py_0
nose=1.3.7=py38_2
notebook=6.1.4=py38_0
numba=0.51.2=py38h0573a6f_1
numexpr=2.7.1=py38h423224d_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
numpydoc=1.1.0=pyhd3eb1b0_1
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openpyxl=3.0.5=py_0
openssl=1.1.1h=h7b6447c_0
opt_einsum=3.1.0=py_0
packaging=20.4=py_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py38h06a4308_1
pango=1.45.3=hd140c19_0
parso=0.7.0=py_0
partd=1.1.0=py_0
patchelf=0.12=he6710b0_0
path=15.0.0=py38_0
path.py=12.5.0=0
pathlib2=2.3.5=py38_0
pathtools=0.1.2=py_1
patsy=0.5.1=py38_0
pcre=8.44=he6710b0_0
pep8=1.7.1=py38_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38he98fc37_0
pip=20.2.4=py38h06a4308_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.14.0=py38h2531618_1
psutil=5.7.2=py38h7b6447c_0
ptyprocess=0.6.0=py38_0
py=1.9.0=py_0
py-lief=0.10.1=py38h403a769_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.6.0=py_0
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.20=py_2
pycurl=7.43.0.6=py38h1ba5d50_0
pydocstyle=5.1.1=py_0
pyflakes=2.2.0=py_0
pygments=2.7.2=pyhd3eb1b0_0
pyjwt=1.7.1=py38_0
pylint=2.6.0=py38_0
pyodbc=4.0.30=py38he6710b0_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38_0
pytables=3.6.1=py38h9fd0a39_0
pytest=6.1.1=py38_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.35.1=py_0
python-libarchive-c=2.9=py_0
pytorch=1.7.1=py3.8_cpu_0
pytz=2020.1=py_0
pywavelets=1.1.1=py38h7b6447c_2
pyxdg=0.27=pyhd3eb1b0_0
pyyaml=5.3.1=py38h7b6447c_1
pyzmq=19.0.2=py38he6710b0_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=h5867ecd_1
qtawesome=1.0.1=py_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
regex=2020.10.15=py38h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
ruamel_yaml=0.15.87=py38h7b6447c_1
scikit-image=0.17.2=py38hdf5156a_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
seaborn=0.11.0=py_0
setuptools=50.3.1=py38h06a4308_1
simplegeneric=0.8.1=py38_2
singledispatch=3.4.0.3=py_1001
termcolor=1.1.0=py38h06a4308_1
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.11.1=py_0
torchvision=0.8.2=py38_cpu
tornado=6.0.4=py38h7b6447c_1
tqdm=4.50.2=py_0
ujson=4.0.1=py38he6710b0_0
unicodecsv=0.14.1=py38_0
unixodbc=2.3.9=h7b6447c_0
wheel=0.35.1=py_0
